Move the disk related operation in src/sys/disk
