+++ src/installer/src/sys/disk/fs/fat12.c
