--- src/installer/src/sys/disk_internal.h
@@ -1,8 +1,0 @@
-#ifndef __SYS__DISK_INTERNAL_H_
-#define __SYS__DISK_INTERNAL_H_
-
-#include <stdint.h>
-
-disk_t *disk_create(uint8_t drive);
-
-#endif
